The-repo's/movie-watchlist-app/requirements.txt
Flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
prometheus-flask-exporter==0.23.0